Feature/board

+++ src/main/java/com/sparta/doguin/domain/board/service/BulletinService.java
@@ -6,6 +6,7 @@ import com.sparta.doguin.domain.board.dto.response.BoardResponse;
 import com.sparta.doguin.domain.board.entity.Board;
 import com.sparta.doguin.domain.board.repository.BoardRepository;
 import com.sparta.doguin.domain.common.exception.HandleNotFound;
+import com.sparta.doguin.domain.common.exception.InvalidRequestException;
 import com.sparta.doguin.domain.common.response.ApiResponseBoardEnum;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -16,13 +17,15 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-public class BulletinService implements BoardService{
+public class BulletinService implements BoardService {
+
     private final BoardRepository boardRepository;
+    private final BoardType boardType = BoardType.BOARD_BULLETIN;
 
     @Override
     @Transactional
-    public Board create(BoardRequest boardRequest,BoardType boardType) {
-        Board board = new Board(boardRequest.title(), boardRequest.content(),boardType);
+    public Board create(BoardRequest boardRequest) {
+        Board board = new Board(boardRequest.title(), boardRequest.content(), boardType);
         return boardRepository.save(board);
     }
 
@@ -31,22 +34,29 @@ public class BulletinService implements BoardService{
     public Board update(Long boardId, BoardRequest boardRequest) {
         Board board = boardRepository.findById(boardId)
                 .orElseThrow(() -> new HandleNotFound(ApiResponseBoardEnum.BULLETIN_NOT_FOUND));
-        board.update(boardRequest.title(),boardRequest.content());
+        if (board.getBoardType() != boardType) {
+            throw new InvalidRequestException(ApiResponseBoardEnum.BULLETIN_WRONG);
+        }
+        board.update(boardRequest.title(), boardRequest.content());
         return board;
     }
 
 
     @Override
     public Board viewOne(Long boardId) {
-        return boardRepository.findById(boardId)
+        Board board = boardRepository.findById(boardId)
                 .orElseThrow(() -> new HandleNotFound(ApiResponseBoardEnum.BULLETIN_NOT_FOUND));
+        if (board.getBoardType() != boardType) {
+            throw new InvalidRequestException(ApiResponseBoardEnum.BULLETIN_WRONG);
+        }
+        return  board;
     }
 
     @Override
-    public Page<BoardResponse> viewAll(int page, int size, BoardType boardType) {
+    public Page<BoardResponse> viewAll(int page, int size) {
         Pageable pageable = PageRequest.of(page - 1, size);
 
-        Page<Board> boards = boardRepository.findAllByBoardType(pageable,boardType);
+        Page<Board> boards = boardRepository.findAllByBoardType(pageable, boardType);
 
         return boards.map(notice -> new BoardResponse(
                 notice.getId(),
@@ -57,9 +67,9 @@ public class BulletinService implements BoardService{
     }
 
     @Override
-    public Page<BoardResponse> search(int page,int size,String title, BoardType boardType) {
+    public Page<BoardResponse> search(int page, int size, String title) {
         Pageable pageable = PageRequest.of(page - 1, size);
-        Page<Board> boards = boardRepository.findAllByTitleAndBoardType(pageable,title,boardType);
+        Page<Board> boards = boardRepository.findAllByTitleAndBoardType(pageable, title, boardType);
 
         return boards.map(notice -> new BoardResponse(
                 notice.getId(),
@@ -72,8 +82,12 @@ public class BulletinService implements BoardService{
     @Transactional
     public void delete(Long boardId) {
 
-        Board board =boardRepository.findById(boardId)
-                .orElseThrow(() -> new HandleNotFound(ApiResponseBoardEnum.NOTICE_NOT_FOUND));
+        Board board = boardRepository.findById(boardId)
+                .orElseThrow(() -> new HandleNotFound(ApiResponseBoardEnum.BULLETIN_NOT_FOUND));
+
+        if (board.getBoardType() != boardType) {
+            throw new InvalidRequestException(ApiResponseBoardEnum.BULLETIN_WRONG);
+        }
 
         boardRepository.delete(board);
     }
